requirements.txt
django==4.2.14
djangorestframework==3.15.2
kafka-python==2.0.2
django-environ==0.11.2
mysqlclient==2.2.4
uwsgi==2.0.26
drf-yasg==1.21.7
